+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 This file contains one line summaries of commits that are worthy of mentioning in release notes.
 A new header is inserted each time a *tag* is created.
 
+- Add a live material debugger with a web-based interface.
+
 ## v1.3.1
 
 - Unified Filament Sceneform and npm releases.